Improve the default error message if an invalid error object is passed to span.setError (#424)
This PR is following a support request where a customer was unable to determine the reason they were seeing a "No message given" default message regularly in their app. This message normally happens when the argument passed to span.setError() doesn't have a .message property, but it doesn't give the customer (or us) a way to find out what the cause of that might be.

This PR improves the message, and attempts to coerce the argument passed to setError to a string and interpolates that into the message.