Add mono to project
Replace lerna with mono, our own tool for managing mono repos.
https://github.com/appsignal/mono/
Follow the instructions on the mono README to install it locally.

Mono is also installed on the CI using the `script/setup` script
which is run before every job. The mono download is cached between
jobs so we don't download it for every job fresh. Update the version
variable `MONO_VERSION` to update the mono install on the CI.
This `setup` script will check the existence of the `cache` command, if
not it will skip it. This allows it to be reused in the test-setups
repo.

Mono depends on yarn workspaces, which are available since yarn 1.