Catch JSON.stringify circular structure error
When a PromiseRejectionEvent has a reason (error object) that contains a
circular structure the `JSON.stringify` function will throw an error.
This break the app because of something we don't handle, so let's catch
the error instead.

First reported in https://github.com/appsignal/support/issues/149

I moved the error reason handling to its own function so it's easier to
read than a (single line) ternary statement. This also makes it easier
to wrap only the `JSON.stringify` in a try-catch-statement.