Add callback argument to sendError and wrap
Currently the `sendError` functions (and in turn the `wrap` helper) has
`tags` and `namespace` parameters. We've seen in the past that this is
quite limiting for the `sendError` helper as users can't add additional
metadata like an action name, parameters, etc.

Deprecate the `tags` and `namespace` parameters in favor of the new
callback function. The deprecations will log a warning to the console
when it's detected they are used. I've not disabled their functionality
so users can upgrade at their own pace.

I'm not very familiar with TypeScript definitions so I've done my best
with what I saw in the rest of the library and found in the docs. The
`tags` parameter in the `send` and `sendError` functions now serves the
double purpose of `tags` parameter and callback function. I've done my
best to differentiate between these two function signatures by defining
multiple signatures.

The thing that was most unclear to me was passing the multi purpose
parameter as a HashMap to the `setTags` function. I ended up casting it
to a HashMap with the `toHashMap` function I saw used elsewhere, only
here I also needed to cast it explicitly to a `HashMap`, hence the
import.

Part of https://github.com/appsignal/integration-guide/issues/41
We also need to update the docs.