Fix version file update on publish
When we publish the JavaScript packages the `version.ts` file in the
JavaScript package wasn't being updated. This was because the
`versionfile` script wasn't run on publish causing the reported version
of the last few releases to be stuck on 1.3.12.

This should work because mono will update the `package.json` before the
build step with the new version. Updating the `version.ts` file after
the build should be fine, could also be done beforehand, it should not
matter.

(The publish step will also call `yarn publish` with the `--new-version
#.#.#` option, and attempt to update the `package.json` with the new
version, but this doesn't do anything as mono already updated the
`package.json` with the new version beforehand.)

Fixes #512