Add build matrix generator
This new build matrix generator is based on the one for Node.js library.
The main difference this one has is that we only test for one Node.js
version (the latest one).

Also added a matrix validation step, and linters for Ruby and JS code.

There's a small modification on test files lookup because React based
libraries files extension ends with "x"