Fix serialization of circular structures of values (#525)
In the plugin-breadcrumbs-console if a user tried to log an object that
was self referencing, that had a circular structure, it would throw an
error that we did not catch and caused an error for the user.

This applies the same fix as for #490: catch the error and send an
alternative value instead.

The console.error message on failure will create another breadcrumb, and
this is fine. This way users know something went wrong with the
breadcrumb.

The console.error on failure to serialize the value will not cause an
infinite loop of breadcrumbs being created. It will only add the error
message as a breadcrumb once.