Rename apiKey to pushApiKey in Node.js installer
In PR https://github.com/appsignal/appsignal-nodejs/pull/507
on the Node.js integration we renamed `apiKey` to `pushApiKey` to make
it more consistent with the other integrations. Update the code in the
installer to match.

I've only not updated the `validatePushApiKey` function, because that
meant updating the core package too. I think it would make more sense
to remove that function from the core package and use the Node.js
package's validator instead. Now we only use the core package for this
one function and that function is not used by the core package.

[skip changeset] as users won't notice this change, it only makes it
more readable for us.
[skip review] as it's only a variable name change.