Serialise objects with circular references
This commit changes the behaviour of the console breadcrumbs
plugin so that it handles circular references, omitting the
cyclic value from the serialised object, instead of throwing an
error.

Before, an object with a circular reference passed to console.log
would fail to be serialised, resulting in a not very useful
breadcrumb, and triggering a second error to be emitted about the
failure to serialise, which polluted the console and the
breadcrumbs output.

After this change, circular references are serialised by replacing
the cyclic value with a string of the form `[cyclic value: $key]`,
where `$key` is the name of the key under which the value was
previously encountered, or "root object" if the root object is
the referenced value.

Note that this can also replace values that don't form cycles, but
that are merely referenced several times across the JSON. Other
solutions to this problem, such as [Crockford's cycle.js][cycle],
also behave in this manner.

The `circularReplacer` function is based on [an MDN snippet][mdn]
about circular references in JSON serialization.

[cycle]: https://github.com/douglascrockford/JSON-js/blob/8e8b0407e475e35942f7e9461dab81929fcc7321/cycle.js#L29-L32
[mdn]: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Errors/Cyclic_object_value#circular_references