Check if thrown object is error in ErrorBoundary
In JavaScript an app can throw any object, not just errors. For example:
`throw new Event("my event")`.

Before we track those thrown objects as errors in AppSignal, make sure
the object is actually an error. I've added the `isError` function for
that in the core library to make this reusable in the future. This
function is then used in the React ErrorBoundary.

For the `isError` tests I had to cast the arguments to `is any`, because
otherwise TypeScript would complain about giving an incompatible type.
Something which the ErrorBoundary itself has no issues with
apparently.

Fixes #532