Serialise objects with circular references
This commit changes the behaviour of the window events plugin so
that it handles circular references, omitting the cyclic value
from the serialised object, instead of throwing an error.

This commit applies the changes to the console breadcrumbs plugin
in [#534][pr] (commit [af829d6][commit]) to the window events
plugin. More details about the changes can be found there.

[pr]: https://github.com/appsignal/appsignal-javascript/pull/534
[commit]: https://github.com/appsignal/appsignal-javascript/commit/af829d6b5b6b1a4e1dae528e35f3648421b31efa