Fix dependency installation in CI build
The CI build did not use the local versions of packages if a specific
dependency version was installed by the CI setup. This because it called
`npm install dependency@version` in the package's directory, rather than
the root. We normally keep all the test dependency packages in the root
directory's `package.json`.

The `npm install` command was also wrong. This repo uses `yarn`, so I've
updated that.

There is a problem with Angular 13. It's documented in issue #539. For
now I've added the previous major version Angular 12 to the build
instead.