Include source inline in sourcemaps
As reported in #540, the sourcemaps don't actually include the contents
of the sources. Only a reference to the file in the `src/` directory of
the package, which we do not include in the published package.

Toggle `inlineSources` on, so that the `.map` file includes the source
of the file, rather than reference it. Alternatively, we need to include
the `src` dir, but this seems more optimized for its purpose.

I didn't want to toggle `inlineSourceMap`, because that makes the main
file that's always loaded upon import way larger, because it merges the
distributed file and the sourcemap into one file.

One small catch to this change: it applies to all packages, and requires
a new release of all packages. Which is why there are so many
changesets.