Check if thrown object is error in ErrorBoundary (#543)
In JavaScript an app can throw any object, not just errors. For example:
`throw new Event("my event")`.

Before we track those thrown objects as errors in AppSignal, make sure
the object is actually an error. I've added the `isError` function for
that in the core library to make this reusable in the future. This
function is then used in the React and Preact ErrorBoundary.

Fixes #532