Check if the correct package version is installed
Move the `yarn add` command to install certain versions of packages to
a separate script so it's also possible to verify if the right versions
of the package are installed afterwards. This gives us more confidence
if the tests matrix works as intended.

It skips the `latest` version, because it won't know what version will
actually be installed.

Related to https://github.com/appsignal/appsignal-nodejs/pull/585
[skip changeset]