Update types dependency
It was listed as a devDependencies in one package, and didn't get
updated by mono on publish. This is fixed in
https://github.com/appsignal/mono/pull/45.

Update the package.json to point to the latest released version.

Remove the explicit types dependency in the main `package.json`. It's
included in the workspace and is not necessary to specify here.

[skip changeset]