Create demo CLI command
Implement a new command for `@appsignal/cli` named "demo". It
reuses the existing logic for sending demonstration sample data
in the "install" command.

The logic for spawning the long-lived process that sends the
sample data has been moved out of the install command, so that it
can be reused in the demo command. For the same reason, the logic
for checking if the `@appsignal/nodejs` package is a dependency of
the `package.json` in the current directory has been moved out of
the diagnose command.

Fixes [appsignal-nodejs#618][issue].

[issue]: https://github.com/appsignal/appsignal-nodejs/issues/618