Add react 18.1.0 to build matrix
Test against the latest stable version of React.

To make it work, several packages had to be updated such as
`@testing-library/react` and `typescript`.

The testing library was the main one causing the test fail as React 18
changes how components are rendered, making version 11 incompatible.
That's why the React 17 tests on the build matrix now explicitly define
a previous testing library.

Typescript was also updated as a dependency, version 4.4 now sets
caught variables as `unknown` type if you have it configured as
`strict`. That feature is disabled in this commit until we update how we
handle errors in the library.