Add changeset for React 18 compatibility
We added React 18 compatibility in our tests and for that, we updated
the integration package.json file. This update is also necessary for
users that want to use React 18. This commit adds a changeset so we can
publish those changes to NPM.

Related PR: https://github.com/appsignal/appsignal-javascript/pull/567