Fix diagnose CLI send report options
The `--no-report` option doesn't match what we have documented how our
diagnose tool works:
https://docs.appsignal.com/nodejs/command-line/diagnose.html

This creates problems for people trying to use the documented
`--send-report` and `--no-send-report` options, which we agreed upon how
this CLI should work.

I've updated the implementation to match our docs. The options can be
set or not, and they are passed to the CLI implementation in the Node.js
package.

Part of https://github.com/appsignal/support/issues/204