Extract message from Error rejection
When reporting an unhandled rejection event, if the reason for the
rejection (the value passed to `reject` in the promise callback)
is an instance of `Error`, derive the name, message and backtrace
of the resulting AppSignal error from it. This usually happens
when the rejection is triggered by a `throw` inside an `async`
function.

Otherwise, default to the same logic as before, reporting the
error as an `UnhandledPromiseRejectionError` and serialising the
reason as a JSON to use that as the error message. The message
"No stacktrace available" where the stacktrace should be is
already added by the `getStacktrace` function that is invoked by
`span.setError`, so it does not need to be added here.