Remove stacktrace check from isError helper
For an object to be considered an error among frameworks it's required
to have a `message` attribute and to be an object. Stacktrace
information is usually an extra.

The old `isError()` behaviour discarded objects without stacktraces
making some of them invisible to customers.

This commit changes the check to consider an object as an error if it
has a message attribute.