Dynamically import `https` module in NodeTransport
In Electron, `https` is not available for import. Importing it
dynamically might fix the issue, if Electron does not attempt to
use the NodeTransport.

Bump the module support to `"es2020"`, which allows for dynamic
imports. This affects what module system is accepted, not what is
emitted.

Emit a console warning if the NodeTransport is used in a context
where HTTPS is not available.