Report errors when running the diagnose
In the `checkForAppsignalPackage` function in `utils.ts`, which is
used by both the diagnose and demo commands, check that the package
is not only present in the `package.json` dependencies, but in the
`node_modules` folder. This is an implicit expectation of both
commands.

In the diagnose command, if calling `spawnSync` returns (but not
raises!) an error, log that error and exit. Before attempting to
spawn the command, check for executable permissions to it, and log
a specific error and exit if we do not have execute permissions for
the diagnose command. Lastly, exit the command with the exit status
of the command called by `spawnSync`.