Bump `@testing-library/react` to 13.x
This removes warnings about the use of `ReactDOM.render` instead
of `createRoot` when using React 18 or earlier. Since our development
dependency on React is already set to `^18`, this seems like the
right move, even though we support earlier versions of React.

See the [React Testing Library 13.0.0 changelog][changelog] and
the [React 18 upgrade guide][guide] for context.

[changelog]: https://github.com/testing-library/react-testing-library/releases/tag/v13.0.0
[guide]: https://react.dev/blog/2022/03/08/react-18-upgrade-guide