Add `override` prop to React error boundary
Add an `override` error prop to the React error boundary. This allows
the customer to pass an override function (named such as a simile to
the client's `appsignal.addOverride` function) that receives the span
that is about to be sent to AppSignal, and optionally the error that
triggered it.

This allows the customer to write custom code that modifies the spans
emitted by this error boundary. This prevents issues with the existing
`tags` prop, where tag values that might be expensive to compute are
passed in, triggering a re-render for errors that will likely not
take place.

It is recommended for the override function that is passed to be
memoized, if declared within the component.