Remove error throw in development
When the app is in development mode our library would throw the error.
This is unlike other integrations and not expected behavior to only
occur in development mode.

In integrations/instrumentations of other libraries we need to rethrow
errors, but here, in our own API we should not throw errors. The error
will be ignored in development mode and an info log message will be
printed to the console.