Do not serialise object keys in place
On `toHashMap` and `toHashMapString`, which are used to normalise
objects that are passed as tags and breadcrumb metadata, do not
serialise the object keys in place. The object passed may be part
of the customer's application, and we should not modify it.

Instead, produce a clone of said object with the serialised keys,
and return that.

I have verified that every use of this utility function makes use
of its return value, rather than calling it for its side effects on
the original object.

Fixes #639.