Explain what "development mode" means
If you misconfigure AppSignal by not providing an API key, it emits
a message saying it's in "development mode", which reads like an
unsurprising mode for a development tool to be in.

Clarify that this "development mode" is caused by not providing an
API key to AppSignal, and that it means that it won't be sending any
data to AppSignal.