Concatenate all match groups
When matching on a backtrace line path using `matchPath`, instead of
taking the first group matched by the regular expression, concatenate
all groups in the regular expression.

This allows for more complex subsets of the path to be selected --
for example, dropping only a specific high-cardinality asset build
folder name from the middle of a path.