Remove global flags from ignored errors
In our public documentation for the `ignoreErrors` config option,
we warn against using the `g` flag on regexes, due to this causing
some errors to not be ignored, since we use `test` internally, and
the regex will remember the last position matched:

- https://docs.appsignal.com/front-end/configuration.html

> Do not use a global flag in any regular expressions, as it may
cause unexpected behavior leading to some errors not being ignored.

- https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/RegExp/test#using_test_on_a_regex_with_the_global_flag

> When a regex has the global flag set, `test()` will advance the
`lastIndex` of the regex. [...] Further calls to `test(str)` will
resume searching `str` starting from `lastIndex`.

What we should do is not document this, but fix it. The `g` flag
is meaningless in these regexes, as we look for any match, so
we can safely remove it when storing the regex.