Remove global flags from backtrace matchers
See previous commit 385e9c2, which fixes this issue for the
`ignoreErrors` regexes, for a broader explanation on the issue that
this addresses.

Store the "unglobalified" regular expressions when initialising the
client. Move some defensive programming from when the config option
is used for a span to when the config option is normalised in the
client.