Match on the entire backtrace path line
On the `SAFARI_FF_PATH` regular expressions, match the whole line
from beginning to end, repeating the matching at the beginning of
the line that is done by the `IS_SAFARI_FF` regular expression.

Split the `CHROME_PATH` regular expression into two variants of it,
with and without a function name. When the function name is present,
the path appears in parenthesis, making it easier to delimit it.
Attempt to match the version where the path is in parenthesis first,
then the version where the path stands alone. On both variants, like
with `SAFARI_FF_PATH`, match the whole line from beginning to end.

This fixes a bug where the regular expressions would only match the
last space-separated segment of a path containing spaces.